10-39 17.08.23 fix

+++ src/service/securityDevicesService.ts
@@ -32,15 +32,17 @@ export const securityDevicesService = {
         if (!refreshToken) return 401;
         const payloadArray = await jwtService.getPayloadRefreshToken(refreshToken)
         if (!payloadArray) return 401;
-       // if(deviceId !== payloadArray[0]) return 403
-        // const refreshTokenMetaObject = await securityDevicesRepo.findRefreshTokenMetaByDeviceId(deviceId)
+        const refreshTokenMetaObject = await securityDevicesRepo.findRefreshTokenMetaByDeviceId(deviceId)
+        if(!refreshTokenMetaObject) return 404
+        if(deviceId !== payloadArray[0]) return 403
+        //
         // if(refreshTokenMetaObject?.userId !== payloadArray[2] )  return 403
         //if(refreshTokenMetaObject && new Date (payloadArray[1]).getTime() !== new Date(refreshTokenMetaObject.lastActiveDate).getTime()) return 401
 
         const deleteSuccess = await securityDevicesRepo.deleteDeviceById(deviceId)
         if(deleteSuccess) return 204
         else return 404
+
     }
 }
 
-/*----*/
